Updated the Swift examples and cleanup up the filter operations now that the compiler won't crash on closures in enums. Also fixed a crash in the lookup filters on the Mac due to a recent pull request.

+++ framework/Source/GPUImageLookupFilter.m
@@ -49,7 +49,7 @@ NSString *const kGPUImageLookupFragmentShaderString = SHADER_STRING
  uniform sampler2D inputImageTexture;
  uniform sampler2D inputImageTexture2; // lookup texture
  
- uniform lowp float intensity;
+ uniform float intensity;
  
  void main()
  {
